+++ pele_platform/Adaptive/plop_solvent.py
